fix for osx and linux
